+++ lib/screens/main_screen/service_menu_screen.dart
@@ -1,6 +1,6 @@
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
-import 'package:greenwheel_user_app/constants/urls.dart';
+import 'package:greenwheel_user_app/core/constants/urls.dart';
 import 'package:greenwheel_user_app/main.dart';
 import 'package:greenwheel_user_app/models/menu_item_cart.dart';
 import 'package:greenwheel_user_app/models/service_type.dart';
@@ -50,7 +50,7 @@ class ServiceMenuScreen extends StatefulWidget {
   final bool? isOrder;
   final String? period;
   final bool? isFromTempOrder;
-  final double? availableGcoinAmount;
+  final int? availableGcoinAmount;
   final void Function() callbackFunction;
   final String? orderGuid;
