+++ lib/screens/plan_screen/list_order_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:greenwheel_user_app/constants/service_types.dart';
-import 'package:greenwheel_user_app/constants/urls.dart';
+import 'package:greenwheel_user_app/core/constants/service_types.dart';
+import 'package:greenwheel_user_app/core/constants/urls.dart';
 import 'package:greenwheel_user_app/screens/main_screen/service_main_screen.dart';
 import 'package:greenwheel_user_app/view_models/location.dart';
 import 'package:greenwheel_user_app/view_models/order.dart';
@@ -25,7 +25,7 @@ class ListOrderScreen extends StatelessWidget {
   final void Function() callback;
   final LocationViewModel location;
   final int memberLimit;
-  final double? availableGcoinAmount;
+  final int? availableGcoinAmount;
   final DateTime endDate;
 
   @override
